Re-introduces legacy default network label selector
This patch uses the legacy label used by CAPW to identify the default
network. This legacy label is used as a fallback in case we cannot find
a default network with the newer CAPV label.
