+++ pkg/services/interfaces.go
@@ -74,12 +74,12 @@ type NetworkProvider interface {
 	// Returns an empty string if the operation is not supported
 	GetClusterNetworkName(ctx *vmware.ClusterContext) (string, error)
 
-	// GetServiceAnnotations returns the annotations, if any, to place on a VM Service.
+	// GetVMServiceAnnotations returns the annotations, if any, to place on a VM Service.
 	GetVMServiceAnnotations(ctx *vmware.ClusterContext) (map[string]string, error)
 
 	// ConfigureVirtualMachine configures a VM for the particular network
 	ConfigureVirtualMachine(ctx *vmware.ClusterContext, vm *vmoprv1.VirtualMachine) error
 
-	// Verify the status of the network after vnet creation
+	// VerifyNetworkStatus verifies the status of the network after vnet creation
 	VerifyNetworkStatus(ctx *vmware.ClusterContext, obj runtime.Object) error
 }
